change language enum type

+++ GW2EIParser/Parser/ParsedData/CombatEvents/MetaDataEvents/LanguageEvent.cs
@@ -2,13 +2,13 @@
 {
     public class LanguageEvent : AbstractMetaDataEvent
     {
-        public enum LanguageEnum : ulong { English = 0, Missing = 1, French = 2, German = 3, Spanish = 4, Unknown = 5 }
+        public enum LanguageEnum : byte { English = 0, Missing = 1, French = 2, German = 3, Spanish = 4, Unknown = 5 }
 
         public LanguageEnum Language { get; }
 
         public LanguageEvent(CombatItem evtcItem) : base(evtcItem)
         {
-            Language = evtcItem.SrcAgent < (ulong)LanguageEnum.Unknown ? (LanguageEnum)evtcItem.SrcAgent
+            Language = evtcItem.SrcAgent < (byte)LanguageEnum.Unknown ? (LanguageEnum)evtcItem.SrcAgent
                 : LanguageEnum.Unknown;
         }
